pyproject.toml
[project]
name = "skuvault-mcp-server"
version = "0.1.0"
description = "MCP server for SkuVault inventory management system"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
skuvault-mcp = "src.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
